代码总体结构调整
分离Equation内部的checkBalance方法，移入新类EquationBalancer；
增加新类：EquationBalancer——方程式配平相关类；
调整.gitignore文件；
