+++ src/org/chembar/glockchem/core/Equation.java
@@ -1,9 +1,7 @@
 package org.chembar.glockchem.core;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 /** 化学方程式抽象
  * @author DuckSoft
@@ -25,59 +23,7 @@ public class Equation {
 	public Equation(String strEquation) throws Exception {
 		this.parseEquation(strEquation);
 	}
-	
-	/** 检查方程式是否平衡
-	 * <p>遍历方程式内{@link Formula}内的原子个数表，以验证方程式是否平衡。<br>
-	 * 若平衡，则返回{@code true}；若否，则返回{@code false}。</p>
-	 * @return 方程式是否平衡
-	 */
-	public boolean checkBalance() {
-		Map<String,Integer> atomReactant = new HashMap<String,Integer>();
-		Map<String,Integer> atomProduct = new HashMap<String,Integer>();
-		
-		for (Pair<Formula,Integer> pair : this.reactant) {
-			for (Map.Entry<String,Integer> entry : pair.getL().mapAtomList.entrySet()) {
-				int numTodo = 0;
-				try {
-					numTodo = atomReactant.get(entry.getKey());
-				} catch(Exception e) {
-					
-				}
-				atomReactant.put(entry.getKey(),entry.getValue() * pair.getR() + numTodo);
-			}
-		}
-		
-		for (Pair<Formula,Integer> pair : this.product) {
-			for (Map.Entry<String,Integer> entry : pair.getL().mapAtomList.entrySet()) {
-				int numTodo = 0;
-				try {
-//					System.out.println(pair.getR());
-					numTodo = atomProduct.get(entry.getKey());
-				} catch(Exception e) {
-				
-				}
-				atomProduct.put(entry.getKey(),entry.getValue() * pair.getR() + numTodo);
-			}
-		}
-		
-//		System.out.print(atomReactant);
-//		System.out.print(atomProduct);
-		for (Map.Entry<String,Integer> entry : atomReactant.entrySet()) {
-			if (atomProduct.containsKey(entry.getKey())) {
-				if (atomProduct.get(entry.getKey()) == entry.getValue()) {
-					continue;
-				} else {
-					return false;
-				}
-			} else {
-				return false;
-			}
-		}
-		
-		return true;
-	}
-	
-	
+
 	/** 分析方程式
 	 * @param strEquation 欲分析的方程式
 	 * @throws Exception
