+++ src/org/chembar/glockchem/core/EquationCalculator.java
